+++ src/index.js
@@ -8,6 +8,23 @@ const ENCODE = [
   ['>', '&gt;'],
 ];
 
+const DATA_KEY = 'hypernova-key';
+const DATA_ID = 'hypernova-id';
+
+// https://gist.github.com/jed/982883
+function uuid() {
+  return (
+    [1e7] +
+    -1e3 +
+    -4e3 +
+    -8e3 +
+    -1e11
+  ).replace(
+    /[018]/g,
+    x => (x ^ Math.random() * 16 >> x / 4).toString(16) // eslint-disable-line no-mixed-operators
+  );
+}
+
 function encode(obj) {
   return ENCODE.reduce((str, coding) => {
     const [encodeChar, htmlEntity] = coding;
@@ -24,35 +41,51 @@ function decode(res) {
   return JSON.parse(jsonPayload);
 }
 
-function backwardCompat(node, attr, key) {
-  return `${node}[data-mystique-key="${key}"],${node}[data-${attr}="${key}"]`;
+function makeValidDataAttribute(attr, value) {
+  const encodedAttr = attr.toLowerCase().replace(/[^0-9a-z_\-]/g, '');
+  const encodedValue = value.replace(/&/g, '&amp;').replace(/"/g, '&quot;');
+  return `data-${encodedAttr}="${encodedValue}"`;
 }
 
-function toScript(attr, key, props) {
-  return `<script type="application/json" data-${attr}="${key}">${LEFT}${encode(props)}${RIGHT}</script>`; // eslint-disable-line max-len
+function toScript(attrs, data) {
+  const dataAttributes = Object.keys(attrs).map(name => makeValidDataAttribute(name, attrs[name]));
+  return `<script type="application/json" ${dataAttributes.join(' ')}>${LEFT}${encode(data)}${RIGHT}</script>`; // eslint-disable-line max-len
 }
 
-function fromScript(attr, key) {
-  const node = document.querySelector(backwardCompat('script', attr, key));
+function fromScript(attrs) {
+  const selectors = Object.keys(attrs)
+    .map(name => `[${makeValidDataAttribute(name, attrs[name])}]`)
+    .join('');
+  const node = document.querySelector(`script${selectors}`);
   if (!node) return null;
   const jsonPayload = node.innerHTML;
+
   return decode(jsonPayload.slice(LEFT.length, jsonPayload.length - RIGHT.length));
 }
 
 function serialize(name, html, data) {
   const key = name.replace(/\W/g, '');
-  const markup = `<div data-hypernova-key="${key}">${html}</div>`;
-  const script = toScript('hypernova-key', key, data);
+  const id = uuid();
+  const markup = `<div data-${DATA_KEY}="${key}" data-${DATA_ID}="${id}">${html}</div>`;
+  const script = toScript({
+    [DATA_KEY]: key,
+    [DATA_ID]: id,
+  }, data);
   return `${markup}\n${script}`;
 }
 
 function load(name) {
   const key = name.replace(/\W/g, '');
-  const node = document.querySelector(backwardCompat('div', 'hypernova-key', key));
-  if (!node) return {};
+  const nodes = document.querySelectorAll(`div[data-${DATA_KEY}="${key}"]`);
 
-  const data = fromScript('hypernova-key', key);
-  return { node, data };
+  return Array.prototype.map.call(nodes, (node) => {
+    const id = node.getAttribute(`data-${DATA_ID}`);
+    const data = fromScript({
+      [DATA_KEY]: key,
+      [DATA_ID]: id,
+    });
+    return { node, data };
+  });
 }
 
 export default function hypernova(runner) {