+++ test/escape-test.js
@@ -15,12 +15,14 @@ describe('escaping', () => {
   .withGlobal('document', () => ({}))
   .describe('with fromScript', () => {
     it('loads the escaped content correctly', () => {
-      const html = toScript('a', 'b', { foo: '</script>', bar: '&gt;', baz: '&amp;' });
+      const html = toScript({ a: 'b' }, { foo: '</script>', bar: '&gt;', baz: '&amp;' });
       const $ = cheerio.load(html);
 
       global.document.querySelector = () => ({ innerHTML: $($('script')[0]).html() });
 
-      const res = fromScript('a', 'b');
+      const res = fromScript({
+        a: 'b',
+      });
 
       assert.isObject(res);
 
@@ -30,7 +32,7 @@ describe('escaping', () => {
     });
 
     it('escapes multiple times the same, with interleaved decoding', () => {
-      const makeHTML = () => toScript('attr', 'key', {
+      const makeHTML = () => toScript({ attr: 'key' }, {
         props: 'yay',
         needsEncoding: '" &gt; </script>', // "needsEncoding" is necessary
       });
@@ -42,7 +44,7 @@ describe('escaping', () => {
 
       global.document.querySelector = () => ({ innerHTML: $($('script')[0]).html() });
 
-      const res = fromScript('attr', 'key');
+      const res = fromScript({ attr: 'key' });
 
       const script3 = makeHTML();
       assert.equal(
@@ -55,5 +57,33 @@ describe('escaping', () => {
 
       assert.equal(res.props, 'yay');
     });
+  });
+
+  it('escapes quotes and fixes data attributes', () => {
+    const markup = toScript({
+      'ZOMG-ok': 'yes',
+      'data-x': 'y',
+      '1337!!!': 'w00t',
+      '---valid': '',
+      'Is this ok?': '',
+      'weird-values': '"]<script>alert(1);</script>',
+      'weird-values2': '"&quot;"',
+    }, {});
+
+    const $ = cheerio.load(markup);
+    const $node = $('script');
+
+    assert.isString($node.data('zomg-ok'));
+    assert.isString($node.data('data-x'));
+    assert.isString($node.data('1337'));
+    assert.isString($node.data('---valid'));
+    assert.isString($node.data('isthisok'));
+
+    assert.equal($node.data('weird-values'), '"]<script>alert(1);</script>');
+    assert.equal($node.data('weird-values2'), '"&quot;"');
+
+    assert.isUndefined($node.data('ZOMG-ok'));
+    assert.isUndefined($node.data('x'));
+    assert.isUndefined($node.data('Is this ok?'));
   });
 });